+++ Disp32/Debug.h
@@ -11,7 +11,7 @@
 const String vFormat (const char *const zcFormat, ...);
 const String vsFormat (const char *const zcFormat, va_list args);
 
-void DebugMsg  (const char* label, int index, const char *const fmt, ...);
-void DebugMsgN (const char* label, int index, String name,  const char *const fmt, ...);
-void DebugMsgF (const char* label, int index, String name, char* flag, const char *const fmt, ...);
+void DebugMsg (const char* label, uint8_t index, const char *const fmt, ...);
+void DebugMsgN (const char* label, uint8_t index, String name,  const char *const fmt, ...);
+void DebugMsgF (const char* label, uint8_t index, String name, char* flag, const char *const fmt, ...);
